Do not install extra header files not used in base build
