+++ service/src/MSRFieldControl.cpp
@@ -33,8 +33,8 @@
 #include "config.h"
 
 #include "MSRFieldControl.hpp"
-#include "geopm/MSRIO.hpp"
-#include "geopm/MSR.hpp"
+#include "MSRIO.hpp"
+#include "MSR.hpp"
 #include "geopm/Exception.hpp"
 #include "geopm_debug.hpp"
