+++ service/test/MSRFieldControlTest.cpp
@@ -38,7 +38,7 @@
 #include "gmock/gmock.h"
 
 #include "MSRFieldControl.hpp"
-#include "geopm/MSR.hpp"
+#include "MSR.hpp"
 #include "geopm/Helper.hpp"
 #include "MockMSRIO.hpp"
 #include "geopm_test.hpp"
